refact : contract ctrl/model 구별
